+++ src/Plivo/Resources/Message/Message.php
@@ -22,6 +22,7 @@ use Plivo\Resources\Resource;
  * @property string $units
  * @property ?string $errorCode
  * @property ?string $powerpackID
+ * @property ?string $requesterIP
  */
 class Message extends Resource
 {
@@ -56,6 +57,13 @@ class Message extends Resource
         }
         if (!empty($response['error_code'])) {
             $this->properties['errorCode'] = $response['error_code'];
+        }
+        if (!empty($response['message_expiry'])) {
+            $this->properties['messageExpiry'] = $response['message_expiry'];
+        }
+
+        if (!empty($response['requester_ip'])) {
+            $this->properties['requesterIP'] = $response['requester_ip'];
         }
 
         $this->pathParams = [