+++ tests/Resources/CampaignTest.php
@@ -75,6 +75,38 @@ class CampaignTest extends BaseTestCase {
         $this->assertRequest($request);
 
     }
+
+    public function testCampaignUpdate()
+    {
+        $campaignID = "CMPT4EP";
+        $request = new PlivoRequest(
+            'POST',
+            'Account/MAXXXXXXXXXXXXXXXXXX/10dlc/Campaign/'.$campaignID.'/',
+            [
+                
+                'reseller_id' => "",
+                'description' => "",
+                'sample1' => "sample 1 should be 20 minimum character",
+                'sample2' => "",
+                'message_flow' => "",
+                'help_message' => "",
+                'optout_message' => "",
+                'optin_keywords' => "",
+                'optout_keywords' => "",
+                'optin_message' => "",
+                'help_keywords' => "",
+                
+            ]);
+        $body = file_get_contents(__DIR__ . '/../Mocks/campaignUpdateResponse.json');
+
+        $this->mock(new PlivoResponse($request,200, $body));
+
+        $actual = $this->client->campaign->update($campaignID, "","","sample 1 should be 20 minimum character","","","","","","","","");
+
+        $this->assertRequest($request);
+
+        self::assertNotNull($actual);
+    }
    
     
     function testCampaignList()
@@ -128,6 +160,23 @@ class CampaignTest extends BaseTestCase {
         $this->assertRequest($request);
         
         self::assertNotNull($actual);
+
+    }
+
+    public function testDeleteCampaign()
+    {
+        $campaignID = "CMPT4EP";
+        $request = new PlivoRequest(
+            'DELETE',
+            'Account/MAXXXXXXXXXXXXXXXXXX/10dlc/Campaign/'.$campaignID.'/',
+            []);
+        $body = file_get_contents(__DIR__ . '/../Mocks/campaignDeleteResponse.json');
+
+        $this->mock(new PlivoResponse($request,200, $body));
+
+        $actual = $this->client->campaign->delete($campaignID);
+        self::assertNotNull($actual);
+        $this->assertRequest($request);
 
     }